modify behavior of holdUnlock
Always increase the channel count during this call even if the wallet
is currently locked.  Additionally, allow the global manager keys to
go from locked to unlocked state while any of these holds are held (by
only requiring the read mutex to unlock).

This will be necessary to not prematurely error for a globally locked
address manager when only individually-encrypted account keys are
needed to perform some function.  Locked errors would be deferred
until attempting to access the locked keys.